requirements.txt
fastapi==0.111.1
langchain_chroma==0.1.2
langchain_core==0.2.19
langchain_openai==0.1.16
pydantic==2.8.2
pymongo==4.8.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.36.0
toml==0.10.2
unstructured==0.14.10
